requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
ruff
google-genai
openai
python-docx
pypdf